+++ src/main/java/ch/ethz/matsim/av/config/operator/OperatorConfig.java
@@ -12,11 +12,13 @@ public class OperatorConfig extends ReflectiveConfigGroup {
 	static public final String ID = "id";
 	static public final String ROUTER_TYPE = "routerType";
 	static public final String ALLOWED_LINK_ATTRIBUTE = "allowedLinkAttribute";
+	static public final String PREDICT_ROUTE_TRAVEL_TIME = "predictRouteTravelTime";
 
 	static public final Id<AVOperator> DEFAULT_OPERATOR_ID = AVOperator.createId("default");
 	private Id<AVOperator> id = DEFAULT_OPERATOR_ID;
 
 	private String allowedLinkAttribute = null;
+	private boolean predictRouteTravelTime = false;
 
 	private final DispatcherConfig dispatcherConfig = new DispatcherConfig();
 	private final GeneratorConfig generatorConfig = new GeneratorConfig();
@@ -64,6 +66,16 @@ public class OperatorConfig extends ReflectiveConfigGroup {
 	@StringSetter(ALLOWED_LINK_ATTRIBUTE)
 	public void setAllowedLinkAttribute(String allowedLinkAttribute) {
 		this.allowedLinkAttribute = allowedLinkAttribute;
+	}
+
+	@StringGetter(PREDICT_ROUTE_TRAVEL_TIME)
+	public boolean getPredictRouteTravelTime() {
+		return predictRouteTravelTime;
+	}
+
+	@StringSetter(PREDICT_ROUTE_TRAVEL_TIME)
+	public void setPredictRouteTravelTime(boolean predictRouteTravelTime) {
+		this.predictRouteTravelTime = predictRouteTravelTime;
 	}
 
 	public DispatcherConfig getDispatcherConfig() {