+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
+- Optionally preroute any AV trip to predict travel times
 - BC: Make vehicle type configurable per operator
 - BC: Add waiting time estimation functionality
 - BC: Add filters for operating areas and use distinct network for each operator